Revert "Also delete the page information cache when invalidating cache of a non-existing page"
Breaks LiquidThreads per bug 43693 and worsens bug 37209.

This reverts commit b3ff701cba6578ff0994ef3b42efcb1ba4eac9cd

+++ includes/Title.php
@@ -4476,8 +4476,11 @@ class Title {
 		HTMLFileCache::clearFileCache( $this );
 
 		// Clear page info.
-		$memcKey = wfMemcKey( 'infoaction', $this->getPrefixedText(), $this->getLatestRevID() );
-		$success = $success && $wgMemc->delete( $memcKey );
+		$revision = WikiPage::factory( $this )->getRevision();
+		if( $revision !== null ) {
+			$memcKey = wfMemcKey( 'infoaction', $this->getPrefixedText(), $revision->getId() );
+			$success = $success && $wgMemc->delete( $memcKey );
+		}
 
 		return $success;
 	}
